requirements.txt
# core requirements
openai==1.14.1
watchdog==4.0.0
halo==0.0.31
inquirer==3.2.4
Jinja2==3.1.3
PyYAML==6.0.1

# user/additional requirements
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
pydantic==2.6.4
